Add files via upload
Every day, tens of thousands of consumers shop at supermarkets for groceries and household goods.

Depending on region or gender, we can use this data to understand customer preferences for payment methods, peak times for visits, and participation in loyalty programs. And we can use this to adjust grocery store policies to raise sales and customer satisfaction.

This project uses a dataset that contains information gathered over three months from three supermarket stores. We can answer a range of questions by examining this dataset.

Does the percentage of customers participating in the rewards program affect gross income or the mode of payment?
Which branch has the best performance in the rewards program?
Does customer feedback affect membership?
Are there any indicators that differ for men and women?